py/objstr: format: Return bytes result for bytes format string.
This is an improvement over previous behavior when str was returned for
both str and bytes input format. This is also consiste how % operator and
many other str/bytes methods work.

It should be noted that it's not how current versions of CPython work,
where there's a gap in the  functionality and bytes.format() is not
supported.